requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blis==1.3.0
breadability==0.1.20
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.2.0
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fonttools==4.54.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.13
mysqlclient==2.2.7
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.2
openai==1.85.0
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2024.8.26
pillow==11.0.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pycountry==24.6.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pynput==1.7.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==80.8.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
sumy==0.11.0
textblob==0.19.0
theHarvester==0.0.1
thinc==8.3.6
threadpoolctl==3.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.2.3
volatility3==2.11.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
